Treat `Item#date` as Date instead of a String
This makes for way less conversions to and from date to string. This way
we only do two conversions: one when we create the object we want to
make sure that the date is `Date` object. The other, when we want to
display the item, we want to convert the date object into a
representable string.